requirements.txt
python-amazon-sp-api
